Fix scan range bug DateTime vs Date
In mhc, scan range of a calendar is set by Date type values.
It woks well, if X-SC-Day field is limited to Date type.

However recent commit allows X-SC-Day to include DateTime
value. For example:
  YYYYMMDD/HH:MM
is allowed in X-SC-Day field for setting different start time
with X-SC-Time: field.

In this case, if scan range (...-20160701) is compared with
a DateTime dtstart value (20160701 10:00), dtstart is counted
as out of range.

This commit fix this kind of off-by-one error.